+++ fx-util/src/main/java/com/dua3/fx/util/FxUtil.java
@@ -75,7 +75,7 @@ public final class FxUtil {
         int a = (argb >> 24) & 0xff;
         int r = (argb >> 16) & 0xff;
         int g = (argb >>  8) & 0xff;
-        int b = (argb >>  0) & 0xff;
+        int b = (argb      ) & 0xff;
 
         return Color.color(r, g, b, a);
     }